Refactor of Hello World. Moved actual construction from Parser to the right class. And cleand up the methods used for it.
